requirements.txt
flask-Session
flask~=3.1.0
markdown~=3.7
requests~=2.32.3
numpy~=2.2.4
fastembed~=0.6.0
flashrank
llama-parse~=0.6.4.post1
qdrant_client
python-dotenv~=1.0.1
together~=1.4.6
yfinance~=0.2.58
GNews~=0.4.1
gradio
colorlog~=6.9.0
bs4~=0.0.2
lxml
pymongo[srv]~=4.11.3
yfinance
torch
pandas~=2.2.3
scikit-learn~=1.6.1
beautifulsoup4~=4.13.3
Werkzeug~=3.1.3
country_converter~=1.3